requirements.txt
python-dateutil==2.2
pystmark==0.4.4
psycopg2==2.7.4
SQLAlchemy==1.2.2
celery==4.1.0
flower==0.9.2
pyYAML==3.12
SQLAlchemy-Utils==0.32.21
GeoAlchemy2 ==0.4.0
nose==1.3.7
coverage==4.0.3
boto3==1.4.4
shapely==1.5.17
raven==6.0.0
lxml==3.8.0
xmljson==0.1.9
backoff==1.4.3
Jinja2==2.9
# pygrowup==0.8.2
